refactor(file-size): split the two zero-headroom files at the 1500 ratchet
crates/stella-context/src/store/tests.rs and
crates/stella-observatory/src/tests.rs each sat at EXACTLY 1500 lines with
no entry in scripts/file-size-baseline.txt. The guard fails at n > 1500, so
both passed with zero headroom: the next line added to either one is a hard
gate failure that the baseline is forbidden to absorb (a file crossing the
limit must be split, and no new baseline entry may be added for it).

That is the same failure that reddened main at 9ad88824, and it was found by
a hand-run headroom scan because the guard reports no warning before a file
crosses -- see #2169.

Pure moves, no behaviour change, grouped by topic rather than by line range:

- store/tests.rs -> store/tests/migrations.rs takes the schema-version and
  vN -> vN+1 migration tests as two spans, deliberately leaving
  editing_a_memory_yields_one_live_record behind: it sits between them but is
  a #712 memory-identity test, not a migration test.
- tests.rs -> tests/execution_routes.rs and tests/http_surface.rs take the
  /api/execution* routes and the raw HTTP surface respectively, both reaching
  the shared seeded_workspace fixtures through the parent's use super::*.

  stella-context store/tests.rs  1500 -> 1105
  stella-observatory tests.rs    1500 ->  988

Refs #2169
