+++ crates/stella-observatory/src/tests.rs
@@ -7,6 +7,13 @@ use tempfile::TempDir;
 
 use super::*;
 
+/// The /api/execution* route tests -- a child module so it reaches the
+/// shared `seeded_workspace` fixtures through this file's own `use super::*`.
+mod execution_routes;
+
+/// The raw HTTP surface tests -- a child module for the same reason.
+mod http_surface;
+
 #[test]
 fn host_header_gates_dns_rebinding() {
     let local = [
@@ -347,291 +354,6 @@ fn models_mirror_stats_semantics() {
     assert_eq!(rows[0]["resolve_rate"], 1.0);
     assert_eq!(rows[1]["division"], "off-grid");
     assert_eq!(rows[1]["cost_per_resolved_usd"], serde_json::Value::Null);
-}
-
-#[test]
-fn execution_detail_includes_steps_tools_files() {
-    let ws = seeded_workspace();
-    let response = respond(ws.path(), "/api/execution?id=1");
-    let v: serde_json::Value = serde_json::from_slice(&response.body).unwrap();
-    assert_eq!(v["id"], 1);
-    assert_eq!(v["steps"].as_array().unwrap().len(), 1);
-    assert_eq!(v["tools"].as_array().unwrap().len(), 2);
-    assert_eq!(v["files"][0]["path"], "src/lib.rs");
-    assert_eq!(v["reflection"]["self_rating"], 8);
-    let none = respond(ws.path(), "/api/execution?id=2");
-    let v: serde_json::Value = serde_json::from_slice(&none.body).unwrap();
-    assert_eq!(
-        v["reflection"],
-        serde_json::Value::Null,
-        "unreflected runs stay null"
-    );
-}
-
-/// The execution detail route names the session that owns the turn.
-///
-/// The transcript is a page with an address (`#transcript/<execution>`), so it
-/// is routinely reached by reload or by a pasted link, with no session list in
-/// memory to have come from. Without this field such an arrival can say
-/// neither which session the turn belongs to nor which turns sit either side
-/// of it, and the page's breadcrumb and prev/next controls are dead. NULL is a
-/// real answer — a run recorded before schema v8 stamped `session_id` has no
-/// session — and must survive as `null` rather than as a missing key, because
-/// the page distinguishes "no session" from "field not served".
-#[test]
-fn execution_detail_carries_its_session_id() {
-    let ws = seeded_workspace();
-    let response = respond(ws.path(), "/api/execution?id=1");
-    let v: serde_json::Value = serde_json::from_slice(&response.body).unwrap();
-    assert_eq!(v["session_id"], "ses-1");
-
-    let unstamped = respond(ws.path(), "/api/execution?id=2");
-    let v: serde_json::Value = serde_json::from_slice(&unstamped.body).unwrap();
-    assert!(
-        v.get("session_id").is_some(),
-        "the key must be served even when the value is null"
-    );
-    assert_eq!(v["session_id"], serde_json::Value::Null);
-}
-
-/// The transcript replay (#1461): the journal route folds an execution's
-/// `events` slice into an ordered transcript, drops `text_delta` fragments
-/// (the `text` event is the authoritative answer), and surfaces tool
-/// arguments and outputs — the content the telemetry projections
-/// deliberately do not carry.
-#[test]
-fn execution_journal_replays_transcript_without_deltas() {
-    let ws = seeded_workspace();
-    let response = respond(ws.path(), "/api/execution-journal?id=1");
-    let v: serde_json::Value = serde_json::from_slice(&response.body).unwrap();
-    let types: Vec<&str> = v
-        .as_array()
-        .unwrap()
-        .iter()
-        .map(|e| e["type"].as_str().unwrap())
-        .collect();
-    assert_eq!(
-        types,
-        [
-            "stage",
-            "reasoning",
-            "tool_start",
-            "tool_result",
-            "text",
-            "text"
-        ],
-        "seq order, text_delta excluded"
-    );
-    assert_eq!(v[0]["label"], "build");
-    assert_eq!(v[1]["body"], "plan the edit");
-    assert_eq!(v[2]["name"], "read_file");
-    assert!(v[2]["body"].as_str().unwrap().contains("src/lib.rs"));
-    assert_eq!(v[3]["ok"], true);
-    assert_eq!(v[3]["body"], "fn a() {}");
-    assert_eq!(v[3]["duration_ms"], 12);
-    // Seq 5 is the pre-#1886 spelling (`delta`), seq 6 the current one
-    // (`text`) — the transcript reads both generations of journal row.
-    assert_eq!(v[4]["body"], "added the function");
-    assert_eq!(v[4]["truncated"], false);
-    assert_eq!(v[5]["body"], "and named it well");
-    // No execution 2 events were seeded — an empty transcript, not an error.
-    let none = respond(ws.path(), "/api/execution-journal?id=2");
-    let v: serde_json::Value = serde_json::from_slice(&none.body).unwrap();
-    assert_eq!(v, serde_json::json!([]));
-}
-
-/// `after_seq` (#1476) narrows the transcript to rows newer than the
-/// drawer's last-seen `seq` — the incremental fetch a still-running
-/// execution's poll uses instead of re-downloading the whole transcript
-/// every tick.
-#[test]
-fn execution_journal_after_seq_returns_only_newer_rows() {
-    let ws = seeded_workspace();
-    // Seq 3 is the tool_start row; only tool_result (4) and the two text
-    // rows (5, 6) — all > 3 — should come back.
-    let response = respond(ws.path(), "/api/execution-journal?id=1&after_seq=3");
-    let v: serde_json::Value = serde_json::from_slice(&response.body).unwrap();
-    let seqs: Vec<i64> = v
-        .as_array()
-        .unwrap()
-        .iter()
-        .map(|e| e["seq"].as_i64().unwrap())
-        .collect();
-    assert_eq!(seqs, [4, 5, 6], "only rows with seq > 3 come back");
-    // A cursor past every seeded row degrades to empty, not an error.
-    let none = respond(ws.path(), "/api/execution-journal?id=1&after_seq=99");
-    let v: serde_json::Value = serde_json::from_slice(&none.body).unwrap();
-    assert_eq!(v, serde_json::json!([]));
-    // Unset behaves exactly like `after_seq=0` would be wrong to: seq 0 (the
-    // opening `stage` row) is still included.
-    let all = respond(ws.path(), "/api/execution-journal?id=1&after_seq=-1");
-    let v: serde_json::Value = serde_json::from_slice(&all.body).unwrap();
-    assert_eq!(
-        v.as_array().unwrap().len(),
-        6,
-        "seq 0 survives after_seq=-1"
-    );
-}
-
-/// A body past [`db::JOURNAL_BODY_CLIP`] chars is clipped and flagged in the
-/// default payload, and returned whole under `?full=1` — the elision must
-/// announce itself, never pose as the data.
-#[test]
-fn execution_journal_clips_long_bodies_unless_full() {
-    use crate::db::JOURNAL_BODY_CLIP;
-    let ws = seeded_workspace();
-    let long = "x".repeat(JOURNAL_BODY_CLIP + 100);
-    Connection::open(ws.path().join(".stella/private/store.db"))
-        .unwrap()
-        .execute(
-            "INSERT INTO events (execution_id, seq, event_type, payload)
-             VALUES (2, 0, 'text', ?1)",
-            [serde_json::json!({"type": "text", "delta": long}).to_string()],
-        )
-        .unwrap();
-    let response = respond(ws.path(), "/api/execution-journal?id=2");
-    let v: serde_json::Value = serde_json::from_slice(&response.body).unwrap();
-    assert_eq!(v[0]["truncated"], true);
-    assert!(v[0]["body"].as_str().unwrap().chars().count() <= JOURNAL_BODY_CLIP + 1);
-    let full = respond(ws.path(), "/api/execution-journal?id=2&full=1");
-    let v: serde_json::Value = serde_json::from_slice(&full.body).unwrap();
-    assert_eq!(v[0]["truncated"], false);
-    assert_eq!(v[0]["body"].as_str().unwrap().chars().count(), long.len());
-}
-
-/// The witness for sent-context inspection (#1475): the reconstructed message
-/// array carries the system prompt the receipt recorded, in wire order, and
-/// says so verifiably.
-///
-/// This is the question the dashboard could not answer — the transcript shows
-/// what a run *did*, never what a call was *sent* — and the system prompt is
-/// the part of it that exists nowhere else in the store.
-#[test]
-fn execution_context_rebuilds_the_message_array_a_call_was_sent() {
-    let ws = seeded_workspace();
-    let response = respond(
-        ws.path(),
-        "/api/execution-context?id=1&turn=0&step=1&call_seq=0",
-    );
-    let v: serde_json::Value = serde_json::from_slice(&response.body).unwrap();
-    assert_eq!(v["receipts_recorded"], true, "{v}");
-    // The index the drawer drills from: both calls that shared step 1.
-    let calls = v["calls"].as_array().unwrap();
-    assert_eq!(calls.len(), 2, "{v}");
-    assert_eq!(calls[0]["call_role"], "worker");
-    assert_eq!(calls[0]["estimated_input_tokens"], 203);
-    assert_eq!(calls[1]["call_role"], "overflow_summarizer");
-    // Frame identity is absent for a receipt written with the lifecycle off,
-    // and null must read as "not computed", never as "computed and empty".
-    assert!(calls[0]["compiled_frame_id"].is_null());
-
-    let context = &v["context"];
-    assert_eq!(context["found"], true, "{v}");
-    let messages = context["messages"].as_array().unwrap();
-    let roles: Vec<&str> = messages
-        .iter()
-        .map(|m| m["role"].as_str().unwrap())
-        .collect();
-    assert_eq!(roles, ["system", "user", "assistant", "tool"], "{v}");
-    assert_eq!(
-        messages[0]["body"], "you are careful",
-        "the system prompt the model actually saw"
-    );
-    assert_eq!(messages[1]["body"], "add a function");
-    // One message, several blocks: the assistant text and the tool call it
-    // carried regroup by `message_index`, in manifest order.
-    assert_eq!(
-        messages[2]["body"],
-        "added the function\n→ tool_call c1 read_file {\"path\":\"src/lib.rs\"}"
-    );
-    assert_eq!(messages[3]["body"], "← tool_result c1\nfn a() {}");
-    // Verified means the journal-resolved blocks re-hashed to the digests the
-    // receipt recorded — a fact re-derived here, not a stored verdict.
-    assert_eq!(context["verified"], true, "{v}");
-    assert_eq!(context["unresolved"], serde_json::json!([]));
-    assert_eq!(context["digest_mismatches"], serde_json::json!([]));
-    // Nothing mismatched, so there is no severity to report — and the era
-    // reads as the pre-rewrite one because this fixture's `executions` table
-    // predates the column, which is exactly how a store written by an older
-    // build behaves (#1981).
-    assert_eq!(context["digest_mismatch_severity"], "none");
-    assert_eq!(context["journal_era"], "compaction_unjournaled");
-    // A gap block's preimage is stored locally, so its check is tautological
-    // and is deliberately not reported as evidence.
-    assert!(messages[0]["blocks"][0]["digest_verified"].is_null());
-    assert_eq!(messages[3]["blocks"][0]["digest_verified"], true);
-    assert_eq!(messages[3]["blocks"][0]["cache_zone"], "volatile");
-    assert_eq!(messages[3]["blocks"][0]["token_cost"], 88);
-}
-
-/// A step can hold several model calls, and each was sent its own context.
-/// Addressing one must never fold in the other's blocks.
-#[test]
-fn execution_context_is_scoped_to_one_call_seq_not_the_whole_step() {
-    let ws = seeded_workspace();
-    let response = respond(ws.path(), "/api/execution-context?id=1&step=1&call_seq=1");
-    let v: serde_json::Value = serde_json::from_slice(&response.body).unwrap();
-    let messages = v["context"]["messages"].as_array().unwrap();
-    assert_eq!(messages.len(), 1, "the summarizer's own manifest: {v}");
-    assert_eq!(messages[0]["role"], "system");
-    // Coordinates with no receipt are answered, not errored.
-    let none = respond(ws.path(), "/api/execution-context?id=1&step=9");
-    let v: serde_json::Value = serde_json::from_slice(&none.body).unwrap();
-    assert_eq!(none.status, "200 OK");
-    assert_eq!(v["context"]["found"], false, "{v}");
-    assert!(
-        v["context"]["note"].as_str().unwrap().contains("step 9"),
-        "{v}"
-    );
-}
-
-/// An execution recorded with `context.lifecycle` off has no receipts, and
-/// that is a state with its own answer — the wording `stella inspect` uses —
-/// not a 500 and not an empty array pretending nothing was sent.
-#[test]
-fn execution_context_without_receipts_says_so_instead_of_erroring() {
-    let ws = seeded_workspace();
-    let response = respond(ws.path(), "/api/execution-context?id=2&step=1");
-    let v: serde_json::Value = serde_json::from_slice(&response.body).unwrap();
-    assert_eq!(response.status, "200 OK");
-    assert_eq!(v["receipts_recorded"], false, "{v}");
-    assert_eq!(v["calls"], serde_json::json!([]));
-    assert!(v["context"].is_null());
-    assert!(
-        v["note"]
-            .as_str()
-            .unwrap()
-            .contains("has no recorded receipts"),
-        "{v}"
-    );
-}
-
-/// A reconstructed message obeys the transcript route's clip policy: past
-/// `db::JOURNAL_BODY_CLIP` chars it is cut and flagged, and `?full=1` returns
-/// the bytes. A system prompt is thousands of stable lines, so this is the
-/// common case here rather than the edge one.
-#[test]
-fn execution_context_clips_long_messages_unless_full() {
-    use crate::db::JOURNAL_BODY_CLIP;
-    let ws = seeded_workspace();
-    let long = "x".repeat(JOURNAL_BODY_CLIP + 100);
-    Connection::open(ws.path().join(".stella/private/store.db"))
-        .unwrap()
-        .execute(
-            "UPDATE context_blocks SET content = ?1 WHERE block_id = 'blk_sys'",
-            [long.as_str()],
-        )
-        .unwrap();
-    let clipped = respond(ws.path(), "/api/execution-context?id=1&step=1");
-    let v: serde_json::Value = serde_json::from_slice(&clipped.body).unwrap();
-    assert_eq!(v["context"]["messages"][0]["truncated"], true, "{v}");
-    let full = respond(ws.path(), "/api/execution-context?id=1&step=1&full=1");
-    let v: serde_json::Value = serde_json::from_slice(&full.body).unwrap();
-    assert_eq!(v["context"]["messages"][0]["truncated"], false);
-    assert_eq!(
-        v["context"]["messages"][0]["body"].as_str().unwrap().len(),
-        long.len()
-    );
 }
 
 #[test]
@@ -1263,238 +985,4 @@ fn hub_telemetry_aggregates_scope_and_drills_to_project() {
     assert_eq!(leaf["name"], "proj-one");
     assert_eq!(leaf["calls"], 2);
     assert!((leaf["cost_usd"].as_f64().unwrap() - 0.75).abs() < 1e-9);
-}
-
-#[test]
-fn unknown_route_is_404_and_missing_id_is_400() {
-    let ws = TempDir::new().unwrap();
-    assert_eq!(respond(ws.path(), "/api/nope").status, "404 Not Found");
-    assert_eq!(
-        respond(ws.path(), "/api/execution").status,
-        "400 Bad Request"
-    );
-    assert_eq!(
-        respond(ws.path(), "/api/execution?id=abc").status,
-        "400 Bad Request"
-    );
-}
-
-#[test]
-fn index_and_mark_are_embedded() {
-    let ws = TempDir::new().unwrap();
-    let index = respond(ws.path(), "/");
-    assert_eq!(index.content_type, "text/html; charset=utf-8");
-    // Lowercase deliberately: the comet brand kit writes "stella — observatory"
-    // (docs/brand/README.md — lowercase always), and this pins the masthead to it.
-    assert!(
-        String::from_utf8(index.body)
-            .unwrap()
-            .contains("stella — observatory")
-    );
-    let mark = respond(ws.path(), "/assets/mark.svg");
-    assert_eq!(mark.content_type, "image/svg+xml");
-    // The header lockup needs both cuts; a missing wordmark route would
-    // render as a broken image rather than failing loudly.
-    let wordmark = respond(ws.path(), "/assets/wordmark.svg");
-    assert_eq!(wordmark.content_type, "image/svg+xml");
-    assert!(String::from_utf8(wordmark.body).unwrap().contains("<svg"));
-}
-
-/// The dashboard sends `encodeURIComponent` output, so a scope value with
-/// a space, `&` or `/` arrives escaped. Comparing the still-encoded bytes
-/// against the stored value silently drilled into an empty scope.
-#[test]
-fn query_param_percent_decodes_the_value() {
-    let q = Some("org=acme%20corp&repo=owner%2Fname&workspace=a+b&empty=");
-    assert_eq!(query_param(q, "org").as_deref(), Some("acme corp"));
-    assert_eq!(query_param(q, "repo").as_deref(), Some("owner/name"));
-    assert_eq!(query_param(q, "workspace").as_deref(), Some("a b"));
-    assert_eq!(query_param(q, "empty"), None);
-    assert_eq!(query_param(q, "absent"), None);
-    assert_eq!(query_param(None, "org"), None);
-    // `%20`/`+` decode to a real space: whitespace is a value, not an
-    // absent filter.
-    assert_eq!(query_param(Some("org=%20"), "org").as_deref(), Some(" "));
-    assert_eq!(query_param(Some("org=+"), "org").as_deref(), Some(" "));
-}
-
-/// This parses attacker-reachable bytes, so a malformed escape must be
-/// inert — never a panic, never a dropped byte. Multi-byte UTF-8 split
-/// across escapes must reassemble; invalid bytes become U+FFFD.
-#[test]
-fn percent_decode_tolerates_malformed_escapes() {
-    assert_eq!(percent_decode("100%"), "100%");
-    assert_eq!(percent_decode("%A"), "%A");
-    assert_eq!(percent_decode("%ZZ"), "%ZZ");
-    assert_eq!(percent_decode("a%%20b"), "a% b");
-    // é is two bytes, escaped separately by encodeURIComponent.
-    assert_eq!(percent_decode("caf%C3%A9"), "café");
-    assert_eq!(percent_decode("%ff"), "\u{fffd}");
-    assert_eq!(percent_decode("plain"), "plain");
-}
-
-/// Inspecting a turn is a route, not an overlay.
-///
-/// This asserts structure, not rendering: that the transcript panel and its
-/// `#transcript/<id>` route exist, that both drill paths (the Overview's
-/// execution table and the Sessions turn table) navigate to it, and that the
-/// modal drawer they used to open is gone rather than merely bypassed — a
-/// second, stale way in is how two renderings of the same data drift apart.
-/// Whether the page then *looks* right is not decidable from Rust; that was
-/// verified by driving the served dashboard in a browser.
-#[test]
-fn inspecting_a_turn_is_a_page_not_a_drawer() {
-    for needle in [
-        "data-tab=\"transcript\"",              // the panel
-        "id=\"panel-transcript\"",              // …addressed by its tab
-        "location.hash = \"transcript/\" + id", // the route
-        "goTranscript(+tr.dataset.exec)",       // sessions → turn
-        "goTranscript(+tr.dataset.id)",         // overview → execution
-    ] {
-        assert!(
-            INDEX_HTML.contains(needle),
-            "the transcript page is missing {needle}"
-        );
-    }
-    for gone in ["id=\"drawer\"", "id=\"scrim\"", "openDrawer(", "aria-modal"] {
-        assert!(
-            !INDEX_HTML.contains(gone),
-            "the execution drawer was replaced by the transcript page, but {gone} survives"
-        );
-    }
-}
-
-/// The page must be fully self-contained: any http(s) URL in the HTML
-/// would be an outbound fetch from the user's browser — a phone-home.
-#[test]
-fn dashboard_html_has_no_external_references() {
-    for needle in ["http://", "https://", "//cdn", "@import", "integrity="] {
-        assert!(
-            !INDEX_HTML.contains(needle),
-            "embedded dashboard must not reference {needle}"
-        );
-    }
-}
-
-#[tokio::test]
-async fn serves_over_a_real_socket() {
-    use tokio::io::{AsyncReadExt, AsyncWriteExt};
-    let ws = seeded_workspace();
-    let root = ws.path().to_path_buf();
-    let (tx, rx) = tokio::sync::oneshot::channel();
-    let server = tokio::spawn(async move {
-        let _ = serve(root, 0, move |addr| {
-            let _ = tx.send(addr);
-        })
-        .await;
-    });
-    let addr = rx.await.unwrap();
-    let mut stream = tokio::net::TcpStream::connect(addr).await.unwrap();
-    stream
-        // A loopback Host: the DNS-rebinding gate (host_is_local, its
-        // own unit test above) would 403 a non-local name.
-        .write_all(b"GET /api/overview HTTP/1.1\r\nHost: 127.0.0.1\r\n\r\n")
-        .await
-        .unwrap();
-    let mut body = String::new();
-    stream.read_to_string(&mut body).await.unwrap();
-    assert!(body.starts_with("HTTP/1.1 200 OK"));
-    assert!(body.contains("\"runs\":2"));
-    // The no-phone-home guarantee, enforced by the browser rather than
-    // only by `dashboard_html_has_no_external_references`.
-    assert!(
-        body.contains("\r\nX-Content-Type-Options: nosniff\r\n"),
-        "head was: {body}"
-    );
-    assert!(
-        body.contains(&format!("\r\nContent-Security-Policy: {CSP}\r\n")),
-        "head was: {body}"
-    );
-    server.abort();
-}
-
-/// The CSP must keep the embedded page working: it is one document with
-/// inline script, an inline `<style>`, and inline `style="…"` attributes,
-/// so dropping `'unsafe-inline'` from either directive would silently
-/// break it. Same-origin `/assets/*.svg` needs `img-src 'self'`.
-#[test]
-fn csp_admits_everything_the_embedded_dashboard_actually_uses() {
-    for directive in [
-        "script-src 'self' 'unsafe-inline'",
-        "style-src 'self' 'unsafe-inline'",
-        "img-src 'self'",
-        "connect-src 'self'",
-        "frame-ancestors 'none'",
-    ] {
-        assert!(CSP.contains(directive), "CSP is missing {directive}");
-    }
-    // A header value may not carry a newline — it would split the head.
-    assert!(!CSP.contains('\r') && !CSP.contains('\n'));
-    assert!(INDEX_HTML.contains("<style>") && INDEX_HTML.contains("<script>"));
-    // Every asset the page names must be same-origin *and* actually routed.
-    // Asserted over whichever attribute carries it rather than a fixed
-    // `src="…"`: the mark moved from an `<img src>` to a `<link rel=icon
-    // href>`, which is the same same-origin fetch and the same CSP question,
-    // but silently failed a literal-string check. What matters is that no
-    // reference points off-origin and none points at a 404.
-    let root = tempfile::tempdir().expect("temp workspace");
-    for attr in ["src=\"", "href=\""] {
-        for (_, rest) in INDEX_HTML
-            .match_indices(attr)
-            .map(|(i, m)| (i, &INDEX_HTML[i + m.len()..]))
-        {
-            let target = rest.split('"').next().unwrap_or_default();
-            assert!(
-                target.starts_with('/') || target.starts_with('#'),
-                "the page must reference nothing off-origin, found {target:?}"
-            );
-            if target.starts_with('/') {
-                assert_ne!(
-                    respond(root.path(), target).status,
-                    "404 Not Found",
-                    "the page references {target:?}, which no route serves"
-                );
-            }
-        }
-    }
-}
-
-/// Padding the request line to the 8 KiB head cap used to be served: the
-/// route still parsed out of the truncated path, while every header —
-/// `Host` included — sat past the cap, and an absent Host is allowed. That
-/// handed a rebound page the whole dashboard cross-origin. A head with no
-/// terminator inside the cap is now refused before routing.
-///
-/// Sized to *exactly* the cap so the server consumes every byte the client
-/// wrote: closing a socket with unread bytes still in its receive queue
-/// sends an RST, which would race the client's read of the response.
-#[tokio::test]
-async fn unterminated_request_head_is_refused_not_routed() {
-    use tokio::io::{AsyncReadExt, AsyncWriteExt};
-    let ws = seeded_workspace();
-    let root = ws.path().to_path_buf();
-    let (tx, rx) = tokio::sync::oneshot::channel();
-    let server = tokio::spawn(async move {
-        let _ = serve(root, 0, move |addr| {
-            let _ = tx.send(addr);
-        })
-        .await;
-    });
-    let addr = rx.await.unwrap();
-    let mut stream = tokio::net::TcpStream::connect(addr).await.unwrap();
-    let prefix = "GET /api/overview?pad=";
-    let suffix = " HTTP/1.1\r\n";
-    let pad = "a".repeat(8192 - prefix.len() - suffix.len());
-    let request = format!("{prefix}{pad}{suffix}");
-    assert_eq!(request.len(), 8192, "fills the cap without terminating");
-    stream.write_all(request.as_bytes()).await.unwrap();
-    let mut body = String::new();
-    stream.read_to_string(&mut body).await.unwrap();
-    assert!(
-        body.starts_with("HTTP/1.1 431"),
-        "an unterminated head must be refused, got: {}",
-        body.lines().next().unwrap_or_default()
-    );
-    assert!(!body.contains("\"runs\""), "no telemetry may leak");
-    server.abort();
 }
